UFSLviaIC/requirements.txt
absl-py==0.11.0
alisuretool==0.0.1
astroid==2.4.2
astunparse==1.6.3
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
isort==5.6.4
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
SoundFile==0.10.3.post1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
termcolor==1.1.0
toml==0.10.2
torch==1.4.0+cu92
torchvision==0.5.0+cu92
tqdm==4.54.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
